// backend/src/modules/sale/services/order-parser/types/fabi.types.ts
export interface FabiOrder {
  id?: string
  created_at?: number
  created_by?: string
  updated_at?: number
  updated_by?: string
  deleted?: boolean
  deleted_by?: any
  deleted_at?: any
  tran_id?: string
  origin_tran_id?: string
  tran_no?: string
  tran_date?: number
  currency?: string
  discount_extra?: number
  discount_extra_amount?: number
  employee_id?: string
  employee_name?: string
  start_date?: number
  start_hour?: number
  start_minute?: number
  end_hour?: number
  end_minute?: number
  sale_note?: string
  sale_type?: string
  sale_method?: any
  service_charge?: number
  service_charge_amount?: number
  shift_id?: string
  ship_fee_amount?: number
  source_fb_id?: string
  table_name?: string
  total_amount?: number
  voucher_amount?: number
  voucher_code?: string
  voucher_name?: string
  foodbook_order_id?: string
  source_voucher?: any
  store_uid?: string
  brand_uid?: string
  company_uid?: string
  table_id?: any
  store_id?: any
  brand_id?: string
  company_id?: any
  device_code?: string
  order_type?: string
  area_id?: any
  source_deli?: string
  sync_status?: number
  discount_extra_name?: any
  service_charge_name?: string
  voucher_extra?: number
  vat_amount?: number
  vat_extra?: number
  amount_discount_detail?: number
  amount_discount_price?: number
  amount_origin?: number
  raw_data?: any
  voucher_amount_paid?: number
  state_action_bill?: number
  version_app?: string
  city_uid?: string
  sale_updated_at?: number
  extra_data?: ExtraData
  discount_extra_amount_campaign?: number
  sale_by_source_payment_status?: string
  commission?: number
  commission_amount?: number
  partner_marketing?: number
  partner_marketing_amount?: number
  is_sync_vat?: number
  vat_invoice_number?: any
  vat_invoice_date?: any
  extra_sale?: ExtraSale
  merged_tran_id?: any
  vat_invoice_series?: any
  vat_tax_reverse_amount?: number
  buffet_id?: any
  discount_vat_amount?: number
  sale_detail?: SaleDetail[]
  sale_payment_method?: SalePaymentMethod[]
}

export interface ExtraData {
  Point?: number
  mkt_max?: number
  comm_max?: number
  del_edit?: number
  ship_fee?: number
  ship_lat?: number
  ship_lng?: number
  peo_count?: number
  tip_amount?: number
  vat_option?: number
  excess_cash?: number
  payment_pda?: number
  process_time?: number
  customer_name?: string
  is_print_bill?: number
  split_two_vat?: number
  customer_email?: string
  customer_phone?: string
  deposit_amount?: number
  money_received?: number
  partner_config?: number
  tran_no_partner?: string
  customer_address?: string
  print_price_temp?: number
  customer_birthday?: string
  Membership_Type_Id?: string
  message_merge_table?: string
  Membership_Type_Name?: string
  call_voucher_partner?: number
  message_modify_table?: string
  not_show_partner_bill?: number
  number_print_check_list?: number
  customer_id?: string
  ship_address?: string
  user_ship_name?: string
  user_ship_phone?: string
}

export interface ExtraSale {}

export interface SaleDetail {
  id?: string
  hour?: number
  note?: string
  image?: any
  point?: number
  price?: number
  amount?: number
  minute?: number
  area_id?: any
  deleted?: boolean
  item_id?: string
  tran_id?: string
  tran_no?: string
  unit_id?: string
  city_uid?: string
  combo_id?: any
  discount?: number
  quantity?: number
  shift_id?: string
  table_id?: any
  toppings?: Topping[]
  brand_uid?: string
  buffet_id?: any
  item_name?: string
  parent_id?: any
  price_org?: number
  store_uid?: string
  tran_date?: number
  created_at?: string
  created_by?: string
  deleted_at?: any
  deleted_by?: any
  package_id?: string
  updated_at?: string
  updated_by?: string
  company_uid?: string
  is_eat_with?: number
  source_deli?: string
  version_app?: string
  discount_vat?: number
  item_type_id?: string
  promotion_id?: any
  vat_tax_rate?: number
  item_class_id?: string
  id_sale_detail?: string
  item_type_name?: string
  origin_tran_id?: string
  vat_tax_amount?: number
  voucher_amount?: number
  discount_amount?: number
  item_class_name?: string
  item_id_mapping?: string
  sale_updated_at?: number
  ship_fee_amount?: number
  sale_detail_type?: number
  commission_amount?: number
  state_action_bill?: number
  unit_secondary_id?: any
  amount_all_topping?: number
  quantity_secondary?: number
  discount_vat_amount?: number
  discount_extra_amount?: number
  service_charge_amount?: number
  discount_values_amount?: number
  vat_tax_reverse_amount?: number
  amount_discount_on_price?: number
  partner_marketing_amount?: number
}

export interface Topping {
  hour?: number
  note?: string
  price?: number
  amount?: number
  is_vat?: number
  minute?: number
  item_id?: string
  tran_id?: string
  unit_id?: string
  discount?: number
  quantity?: number
  item_name?: string
  price_org?: number
  package_id?: string
  is_eat_with?: number
  sub_topping?: any[]
  check_active?: number
  discount_vat?: number
  item_type_id?: string
  raw_quantity?: number
  time_cooking?: number
  vat_tax_rate?: number
  item_class_id?: string
  quantity_done?: number
  id_sale_detail?: string
  item_type_name?: string
  split_quantity?: number
  vat_tax_amount?: number
  voucher_amount?: number
  discount_amount?: number
  item_class_name?: string
  item_id_mapping?: string
  ship_fee_amount?: number
  commission_amount?: number
  amount_all_topping?: number
  not_update_quantop?: number
  quantity_secondary?: number
  discount_vat_amount?: number
  quantity_print_temp?: number
  discount_extra_amount?: number
  service_charge_amount?: number
  discount_values_amount?: number
  vat_tax_reverse_amount?: number
  amount_discount_on_price?: number
  partner_marketing_amount?: number
}

export interface SalePaymentMethod {
  id?: string
  image?: any
  amount?: number
  area_id?: any
  deleted?: boolean
  is_paid?: number
  tran_id?: string
  tran_no?: string
  city_uid?: string
  currency?: string
  pay_type?: string
  shift_id?: string
  table_id?: any
  trace_no?: string
  brand_uid?: string
  buffet_id?: any
  store_uid?: string
  tran_date?: number
  created_at?: string
  created_by?: string
  deleted_at?: any
  deleted_by?: any
  updated_at?: string
  updated_by?: string
  company_uid?: string
  order_count?: number
  payment_fee?: number
  source_deli?: string
  version_app?: string
  origin_tran_id?: string
  sale_updated_at?: number
  payment_method_id?: string
  state_action_bill?: number
  payment_fee_amount?: number
  tran_id_of_partner?: any
  payment_method_name?: string
}